Handle EOF condition correctly
Before spawning a new readahead thread we should check
if there's data in the readahead buffer as the readahead
thread may have exited because it reached EOF but there
may still be data in the buffer.

An offset field was also added to CurlUtil_File. It is only
used for debug output at the moment but it will be used in
the future in improve seek time.