Cleanup shutdown process and add ipv4 only support
Cleanup the shutdown process to prevent long delays or
lockups during shutdown and allow compiling against an
ipv4 only build of libupnp. Note that (as of libupnp 1.6.18)
an IPv6 enabled libupnp will not work under a kernel without
IPv6 support.

Also use a relative path for config.h to make sure that
we use the right header on all systems.