Fix issues building on debug versions of libupnp
When libupnp is built with debug support it may define UPNP_HAVE_DEBUG
to 1 and djmount used that to determine if SetLogFileNames is available,
but apparently this function comes comes from a patch to the libupnp
that came bundled with djmount. So for now we'll just check for that
function at configure time and only use it if it's definitely available
(probably never).