Added Ubuntu build instructions and fix warnings
Added build instructions for Ubuntu 16.04 LTS and fixed warnings
when compiling against libupnp 2.6.19 (ships with 16.04). The warnings
are caused patches to ubuntu's libupnp. As of todoy Ubuntu's 16.04
libupnp version is "1.6.19+git20160116" and it has several members
declared as type UpnpString but they're char arrays upstream (on the whole
1.16.x branch). Since we don't have a reliable way to check for Ubuntu's
version we'll check the type of each member individual during configure.