Workaround libupnp/IPv6 issues
Under some configurations the IPv6 enabled UpnpInit2() function may fail
in some cases where the deprecated UpnpInit() function works. One such
scenario is when running under a kernel with no IPv6 support but it may
also be causes by other IPv6 configuration issues. As a workaround we'll
try UpnpInit2() first whenever IPv6 is available but if it fails we'll
just log an error and try the IPv4 only UpnpInit().