Fixed race condition that caused random crash
When a network interface is detected we where firing the client process
and adding it to the list of clients before waiting for it to be ready.
If the client fails to start up (because the network interface is not
really up yet) we'll have a short period when we have an interface entry
on the list but the client already exited. If a request is sent to that
interface we crash since there's no client to handle it.