Move interface enumeration code to reusable file and added IP check.
The interface enumeration code has been moved to a new file and a check
to make sure that the interface has a valid ip has been added. This
should fix the client launch failures that sometimes happen just after
the interface is brought up. This amounts to less logged errors (since
we're successfully recovering) and better IPv6 support (since IPv6 gets
disabled when the client fails to start).