Switch log locking to pthreads instead of ithread
Log_Printf() is sometimes getting stuck on fflush() when built against
uClibc. The first step in diagnosing this issue is to switch from the
ithread mutex (which I'm not familiar with) and use pthreads to rule out
an ithreads issue).

Edit: It turns out the problem was not related to avmount (the process
that starts avmount was setting it's stdout to a pipe but it wasn't
reading it so it blocks once the pipe's buffer fills). However I'm
keeping this commit as it's cleaner than the prior implementation. Also
the ithread implementation uses a recursive mutex unnecessarily.