Just wait when curl_multi_fdset() returns -1
All of the sudden curl_multi_fdset() started returning -1. When this
happens we're sleeping 100ms and trying again, and so it appears that it
was never happening.

The cURL documentations says: "When libcurl returns -1 in max_fd, it is
because libcurl currently does something that isn't possible for your
application to monitor with a socket and unfortunately you can then not
know exactly when the current action is completed using select(). You
then need to wait a while before you proceed and call curl_multi_perform
anyway." That is what we're doing now. I still don't understand why this
just started happening. Aparently it was caused to an update to libcurl
from from 7.54.1 to 7.55.1 but I need to look into that.